move execute to wire on a node
