+++ pkg/workflow/execute/executor.go
@@ -1,60 +1,16 @@
 package execute
 
 import (
-	"fmt"
 	"github.com/pkg/errors"
 	"github.com/protoflow-labs/protoflow/pkg/workflow/graph"
-	"github.com/protoflow-labs/protoflow/pkg/workflow/node"
-	"github.com/reactivex/rxgo/v2"
 	"go.temporal.io/sdk/workflow"
 )
 
-type Input struct {
-	Observable rxgo.Observable
-	Resource   graph.Resource
-}
-
-type Output struct {
-	Observable rxgo.Observable
-}
-
 type Executor interface {
-	Execute(n graph.Node, input Input) (*Output, error)
+	Execute(n graph.Node, input graph.Input) (*graph.Output, error)
 }
 
 var _ Executor = &TemporalExecutor{}
-
-var activity = &Activity{}
-
-// TODO breadchris each node should have an Execute function, however at the moment, an import cycle would be formed by
-// node -> resource -> node. Need to figure out how to avoid this.
-func nodeToActivityName(n graph.Node) ActivityFunc {
-	switch n.(type) {
-	case *node.RESTNode:
-		return activity.ExecuteRestNode
-	case *node.GRPCNode:
-		return activity.ExecuteGRPCNode
-	case *node.FunctionNode:
-		return activity.ExecuteFunctionNode
-	case *node.CollectionNode:
-		return activity.ExecuteCollectionNode
-	case *node.BucketNode:
-		return activity.ExecuteBucketNode
-	case *node.InputNode:
-		return activity.ExecuteInputNode
-	case *node.FileNode:
-		return activity.ExecuteFileNode
-	case *node.QueryNode:
-		return activity.ExecuteQueryNode
-	case *node.PromptNode:
-		return activity.ExecutePromptNode
-	case *node.RouteNode:
-		return activity.ExecuteRouteNode
-	case *node.TemplateNode:
-		return activity.ExecuteTemplateNode
-	}
-	return nil
-}
 
 type TemporalExecutor struct {
 	ctx workflow.Context
@@ -66,13 +22,10 @@ func NewTemporalExecutor(ctx workflow.Context) *TemporalExecutor {
 	}
 }
 
-func (e *TemporalExecutor) Execute(n graph.Node, input Input) (*Output, error) {
-	var result Output
-	act := nodeToActivityName(n)
-	if act == nil {
-		return nil, fmt.Errorf("error getting activity for node: %s", n.NormalizedName())
-	}
-	err := workflow.ExecuteActivity(e.ctx, act, n, input).Get(e.ctx, &result)
+func (e *TemporalExecutor) Execute(n graph.Node, input graph.Input) (*graph.Output, error) {
+	var result graph.Output
+	// TODO breadchris n.Wire will not work here since n is a pointer and we are changing execution context, i think
+	err := workflow.ExecuteActivity(e.ctx, n.Wire, n, input).Get(e.ctx, &result)
 	if err != nil {
 		return nil, errors.Wrap(err, "error executing activity")
 	}
@@ -97,13 +50,8 @@ func NewMemoryExecutor(ctx *MemoryContext, opts ...MemoryExecutorOption) *Memory
 	return e
 }
 
-func (e *MemoryExecutor) Execute(n graph.Node, input Input) (*Output, error) {
-	act := nodeToActivityName(n)
-	if act == nil {
-		return nil, fmt.Errorf("error getting activity for node: %s", n.NormalizedName())
-	}
-
-	res, err := act(e.ctx.Context, n, input)
+func (e *MemoryExecutor) Execute(n graph.Node, input graph.Input) (*graph.Output, error) {
+	res, err := n.Wire(e.ctx.Context, input)
 	if err != nil {
 		return nil, err
 	}
