--- pkg/workflow/execute/static.go
@@ -1,1 +1,0 @@
-package execute
