+++ pkg/workflow/node/base.go
@@ -5,6 +5,7 @@ import (
 	"github.com/protoflow-labs/protoflow/gen"
 	"github.com/protoflow-labs/protoflow/pkg/util"
 	"github.com/protoflow-labs/protoflow/pkg/workflow/graph"
+	"github.com/rs/zerolog/log"
 	"strings"
 )
 
@@ -12,6 +13,15 @@ type BaseNode struct {
 	Name       string
 	id         string
 	resourceID string
+}
+
+// NewBaseNode creates a new BaseNode from a gen.Node, gen.Node cannot be embedded into BaseNode because proto deserialization will fail on the type
+func NewBaseNode(node *gen.Node) BaseNode {
+	return BaseNode{
+		Name:       util.ToTitleCase(node.Name),
+		id:         node.Id,
+		resourceID: node.ResourceId,
+	}
 }
 
 func (n *BaseNode) NormalizedName() string {
@@ -31,65 +41,12 @@ func (n *BaseNode) ResourceID() string {
 }
 
 func (n *BaseNode) Info(r graph.Resource) (*graph.Info, error) {
+	log.Warn().
+		Str("node", n.Name).
+		Msg("Info() not implemented")
 	return nil, nil
 }
 
 func (n *BaseNode) Represent() (string, error) {
 	return "", errors.New("not implemented")
 }
-
-type RESTNode struct {
-	BaseNode
-	*gen.REST
-}
-
-var _ graph.Node = &RESTNode{}
-
-type CollectionNode struct {
-	BaseNode
-	Collection *gen.Collection
-}
-
-var _ graph.Node = &CollectionNode{}
-
-type BucketNode struct {
-	BaseNode
-	*gen.Bucket
-}
-
-var _ graph.Node = &BucketNode{}
-
-type QueryNode struct {
-	BaseNode
-	Query *gen.Query
-}
-
-var _ graph.Node = &QueryNode{}
-
-type PromptNode struct {
-	BaseNode
-	Prompt *gen.Prompt
-}
-
-var _ graph.Node = &PromptNode{}
-
-type SecretNode struct {
-	BaseNode
-	Secret *gen.Secret
-}
-
-var _ graph.Node = &SecretNode{}
-
-type TemplateNode struct {
-	BaseNode
-	Template *gen.Template
-}
-
-var _ graph.Node = &TemplateNode{}
-
-type FileNode struct {
-	BaseNode
-	File *gen.File
-}
-
-var _ graph.Node = &FileNode{}
