+++ pkg/workflow/node/config.go
@@ -1,6 +1,8 @@
 package node
 
 import (
+	"context"
+	"github.com/pkg/errors"
 	"github.com/protoflow-labs/protoflow/gen"
 	"github.com/protoflow-labs/protoflow/pkg/workflow/graph"
 )
@@ -21,4 +23,8 @@ func NewConfigNode(node *gen.Node) *ConfigNode {
 
 func (c *ConfigNode) Represent() (string, error) {
 	return c.Config.Value, nil
+}
+
+func (c *ConfigNode) Wire(ctx context.Context, input graph.Input) (graph.Output, error) {
+	return graph.Output{}, errors.New("not implemented")
 }
