+++ pkg/workflow/node/function.go
@@ -1,10 +1,13 @@
 package node
 
 import (
+	"context"
+	"fmt"
 	"github.com/pkg/errors"
 	"github.com/protoflow-labs/protoflow/gen"
 	"github.com/protoflow-labs/protoflow/pkg/workflow/graph"
 	"github.com/protoflow-labs/protoflow/pkg/workflow/resource"
+	"github.com/rs/zerolog/log"
 	"strings"
 )
 
@@ -20,6 +23,22 @@ func NewFunctionNode(node *gen.Node) *FunctionNode {
 		BaseNode: NewBaseNode(node),
 		Function: node.GetFunction(),
 	}
+}
+
+func (n *FunctionNode) Wire(ctx context.Context, input graph.Input) (graph.Output, error) {
+	log.Debug().
+		Str("name", n.NormalizedName()).
+		Msg("setting up function")
+	g, ok := input.Resource.(*resource.LanguageServiceResource)
+	if !ok {
+		return graph.Output{}, fmt.Errorf("error getting language service resource: %s", n.Name)
+	}
+
+	// provide the grpc resource to the grpc gn call. Is this the best place for this? Should this be provided on injection? Probably.
+	input.Resource = g.GRPCResource
+
+	grpcNode := n.ToGRPC(g)
+	return grpcNode.Wire(ctx, input)
 }
 
 func (n *FunctionNode) Info(r graph.Resource) (*graph.Info, error) {
