+++ pkg/workflow/node/input.go
@@ -1,6 +1,7 @@
 package node
 
 import (
+	"context"
 	"github.com/protoflow-labs/protoflow/gen"
 	"github.com/protoflow-labs/protoflow/pkg/workflow/graph"
 )
@@ -17,4 +18,10 @@ func NewInputNode(node *gen.Node) *InputNode {
 		BaseNode: NewBaseNode(node),
 		Input:    node.GetInput(),
 	}
+}
+
+func (n *InputNode) Wire(ctx context.Context, input graph.Input) (graph.Output, error) {
+	return graph.Output{
+		Observable: input.Observable,
+	}, nil
 }
